Correctly clean up various DOM manipulations
Previously there were certain cases where the DOM was manipulated
(adding body classes etc.) during `beforeOpen` which was not correctly
cleaned up when unmounting. This was due to a race condition between
setting the `isOpen` state and the `componentWillUnmount` lifecycle
method being called.

To fix this, we use a regular instance variable to track when we've
manipulated the DOM and therefore when we need to clean up after
ourselves, rather than relying on state.